Updated documentation (needed for SAVE/SAVE AS), other changes also.
The general DCE document needed to be updated for the changes to the
functioning of SAVE and SAVE AS. Other updates were made to fix a
few minor issues. For example, the word for "WIRE" was scrambled.

Tutorial 1 was also reviewed with minor corrections. It may be that
the wiring for the switches in tutorial 1 is reversed, but it's not
been confirmed by anyone else. At this time, the wiring remains as
it was in the original. It should be checked again carefully.