Added Thread.yield per comments provided by Gaston Williams
Gaston wrote (on COSMAC ELF in groups.io):

Hi Bob,

I believe yield is a new keyword associated with the switch statement.
From googling, it looks like it was introduced with Java 13.  This new
keyword collides with previously named references and methods, in
particular Thread.yield().  So to prevent an ambiguous reference they
now have to be qualified.

Since Dce extends Thread, you can just use this.yield() to invoke the
yield method and get rid of the error message. In other cases, one can
invoke the method with Thread.yield() since it's a public static method
in the Thread class.

Best regards,
Gaston