Added ability to log probed values to the console, and Clock.txt.
To enable logging, use the menu item Set/Log_Probe. This will
toggle the logging on and off. The initial value is off. The
toggled value is printed to the console window along with the
names and values of each pin being probed.

This is a very primitive logging system. When logging is enabled,
it prints the names of the currently logged pins to the console.
As the simulation runs, the logged pin values (0 or 1) are also
printed to the console one line at a time. The resulting output
of the run contains a "header" line of space-separated pin names
followed by a series of space separated values (0 or 1) on each
line for each cycle of the simulation. These can be copied from
the terminal or the output can be redirected to a file.

This capability is easiest seen by simply simulating a clock chip
with all of its pins being probed. The resulting output shows the
various clock waveforms generated by the clock chip.