Added ability to flip data sheets and added 1802 to 1802 menus.
This change adds a "Flip" button to the upper left corner of the
data sheet display window. Clicking the "flip" button will flip
the drawing from right to left to show the bottom side. This is
particularly helpful when wiring from the bottom side in JDCE.

This change also added the CDP1802BC microprocessor to both
of the Parts and Data menus under CDP 1800 Series. The 1802
already exists under the "Miscelaneous Chips" menu, and it is
still there. But it seemed missing from the 1802 menu, so it
was added there as well. It might be removed from the Misc.
menu at some time.