Rearranged order of evaluating action commands to ensure they work.
Some of the action commands are decoded by looking for a single
character at the beginning of the action command (such as "C" or "D").
If the check for other longer commands (like "Cut" or "Delete") is
done later in the order, then the "C" and "D" actions will always
happen and the "Cut" and "Delete" actions will never happen. That
was the case with the "Disable Logging" command which was being
eclipsed by a check for "D" as the first letter. This change just
rearranged the items so that the most unique are tested first, and
the least unique are tested last.