Added support for semicolon comments (;...) in circuit files
This change added a "readLine" method to the "Board" class to read
a file and filter out comments (which are currently discarded). The
new readLine method ("this.readLine(file)") replaces reading from
the file itself ("file.readLine()") in the loadFile method. The new
method filters each line to remove all text following and including
a semicolon (;). It also removes any empty (whitespace only) lines.
This allows the loadFile method to operate as it had in the past.

Note that the use of the semicolon (;) could be easily replaced
with any other preferred character.